Добавлены пользователи и db (1.5/2)
Не протестировано, не доделано.
