+++ backend/orchestrator/config.go
@@ -1,6 +1,10 @@
 package main
 
-import "net/http"
+import (
+	"database/sql"
+	"log"
+	"net/http"
+)
 
 func GetDefaultHttpServer(handler http.Handler) *http.Server {
 	return &http.Server{Addr: "127.0.0.1:8000", Handler: handler}
@@ -8,4 +12,12 @@ func GetDefaultHttpServer(handler http.Handler) *http.Server {
 
 func GetDefaultGrpcServer() *GrpcTaskServer {
 	return &GrpcTaskServer{Addr: "127.0.0.1:5000"}
+}
+
+func GetDefaultSqlServer() *sql.DB {
+	var db, err = sql.Open("sqlite3", "store.db")
+	if err != nil {
+		log.Panic(err)
+	}
+	return db
 }
