+++ backend/utils.go
@@ -2,7 +2,10 @@ package backend
 
 import (
 	"errors"
+	jwt2 "github.com/golang-jwt/jwt/v5"
+	"log"
 	"os"
+	"time"
 )
 
 type EnvVar struct {
@@ -44,5 +47,25 @@ func convertToInt64Interface(arg interface{}) (result interface{}, err error) {
 	default:
 		err = errors.New("arg должен быть числом")
 	}
+	return
+}
+
+const hmacSampleSecret = "test_test"
+
+func GenerateJwt(user main.DbUser) (tokenBuf []byte, err error) {
+	var currentTime = time.Now()
+	var jwtInstance = jwt2.NewWithClaims(jwt2.SigningMethodHS256, jwt2.MapClaims{
+		"login": user.GetLogin(),
+		"id":    user.GetId(),
+		"nbf":   currentTime.Unix(),
+		"exp":   currentTime.Add(5 * time.Minute).Unix(),
+		"iat":   currentTime.Unix(),
+	})
+	var token string
+	token, err = jwtInstance.SignedString([]byte(hmacSampleSecret))
+	if err != nil {
+		log.Panic(err)
+	}
+	tokenBuf = []byte(token)
 	return
 }
